+++ src/app/services/selfhelp.service.ts
@@ -17,6 +17,8 @@ import packageJson from './../../../package.json'; // Replace with the actual pa
 import config from 'capacitor.config';
 import appConfig from 'src/env/app.config';
 import { SavePassword } from 'capacitor-ios-autofill-save-password';
+import { App } from '@capacitor/app';
+import { AppLauncher } from '@capacitor/app-launcher';
 const appVersion = packageJson.version;
 
 
@@ -911,4 +913,15 @@ export class SelfhelpService {
             }
         });
     }
+
+    async openAnotherApp() {
+        const packageName = "https://play.google.com/store/apps/details?id=com.burockgames.timeclocker&hl=en&gl=US"; // Replace with the package name of the app you want to open
+
+        try {
+          await AppLauncher.openUrl({ url: packageName });
+          console.log('Successfully opened the app');
+        } catch (error) {
+          console.error('Error opening the app:', error);
+        }
+      }
 }